Struct lazy_clone_collections::btree::BTreeMap

pub struct BTreeMap<K, V, const N: usize = 2> { /* private fields */ }

Implementations

impl<K, V, const N: usize> BTreeMap<K, V, N>

pub fn append(&mut self, other: &mut Self)where
K: Clone + Ord,
V: Clone,

pub fn clear(&mut self)

pub fn contains_key<Q>(&self, key: &Q) -> boolwhere
K: Borrow<Q>,
Q: Ord + ?Sized,

pub fn entry(&mut self, key: K) -> Entry<'_, Self>where
K: Clone + Ord,
V: Clone,

Returns an Entry that simplifies some update operations.

pub fn first_entry(&mut self, key: K) -> Entry<'_, Self>where
K: Clone + Ord,
V: Clone,

Return an Entry for the least key in the map.

pub fn first_key_value(&self) -> Option<(&K, &V)>

pub fn get<Q>(&self, key: &Q) -> Option<&V>where
K: Borrow<Q>,
Q: Ord + ?Sized,

Retrieves the value associated with the given key, if there is one.

Examples

use lazy_clone_collections::BTreeMap;

let mut m = BTreeMap::new();
m.insert(0, 1);
assert_eq!(m.get(&0), Some(&1));
assert_eq!(m.get(&1), None);

pub fn get_key_value<Q>(&self, key: &Q) -> Option<(&K, &V)>where
K: Borrow<Q>,
Q: Ord + ?Sized,

Retrieves the entry associated with the given key, if there is one.

Examples

use lazy_clone_collections::BTreeMap;

let mut m = BTreeMap::new();
m.insert(0, 1);
assert_eq!(m.get_key_value(&0), Some((&0, &1)));
assert_eq!(m.get(&1), None);

pub fn get_mut<Q>(&mut self, key: &Q) -> Option<&mut V>where
K: Borrow<Q> + Clone,
V: Clone,
Q: Ord + ?Sized,

pub fn insert(&mut self, key: K, val: V) -> Option<V>where
K: Clone + Ord,
V: Clone,

Associates ‘val’ with ‘key’ and returns the value previously associated with ‘key’, if it exists.

Examples

use lazy_clone_collections::BTreeMap;

let mut m = BTreeMap::new();
assert_eq!(m.insert(0, 1), None);
assert_eq!(m.insert(0, 0), Some(1));

pub fn into_keys(self) -> impl Iterator<Item = K>where
K: Clone,
V: Clone,

pub fn into_values(self) -> impl Iterator<Item = V>where
K: Clone,
V: Clone,

pub fn is_empty(&self) -> bool

pub fn iter(&self) -> Iter<'_, K, V, N>

pub fn iter_mut(&mut self) -> IterMut<'_, K, V, N> where
K: Clone,
V: Clone,

pub fn keys(&self) -> impl Iterator<Item = &K>

pub fn last_entry(&mut self, key: K) -> Entry<'_, Self>where
K: Clone + Ord,
V: Clone,

Return an Entry for the least key in the map.

pub fn last_key_value(&self) -> Option<(&K, &V)>

pub fn len(&self) -> usize

pub fn new() -> Self

pub fn pop_first(&mut self) -> Option<(K, V)>where
K: Clone,
V: Clone,

pub fn pop_last(&mut self) -> Option<(K, V)>where
K: Clone,
V: Clone,

pub fn range<Q, R>(&self, range: R) -> Range<'_, K, V, N> where
Q: Ord + ?Sized,
K: Borrow<Q>,
R: RangeBounds<Q>,

pub fn range_mut<Q, R>(&mut self, range: R) -> RangeMut<'_, K, V, N> where
Q: Ord + ?Sized,
K: Borrow<Q> + Clone,
R: RangeBounds<Q>,
V: Clone,

pub fn remove<Q>(&mut self, key: &Q) -> Option<V>where
K: Borrow<Q> + Clone,
V: Clone,
Q: Ord + ?Sized,

Removes and returns the value associated with key, if it exists.

Examples

use lazy_clone_collections::BTreeMap;

let mut m = BTreeMap::new();
m.insert(0, 'a');
m.insert(1, 'b');
assert_eq!(m.remove(&0), Some('a'));
assert_eq!(m.remove(&1), Some('b'));
assert_eq!(m.remove(&1), None);
assert_eq!(m.len(), 0);

pub fn retain<F>(&mut self, f: F)where
K: Clone + Ord,
V: Clone,
F: FnMut(&K, &mut V) -> bool,

pub fn split_off<Q>(&mut self, key: &Q) -> Selfwhere
Q: Ord + ?Sized,
K: Borrow<Q> + Clone + Ord,
V: Clone,

pub fn values(&self) -> impl Iterator<Item = &V>

pub fn values_mut(&mut self) -> impl Iterator<Item = &mut V>where
K: Clone,
V: Clone,