Struct lazy_clone_collections::AvlMap

pub struct AvlMap<K, V> { /* private fields */ }

A map from keys to values sorted by key.

We aim for the API to be compatible with the (stable) API of std::collections::BTreeMap.

Internally, the map uses ‘persistent’ AVL trees. AVL trees were the first self-balancing binary tree. The trees are persistent in that trees cloned from a common ancestor will share nodes until they are updated. When a map is updated, it creates a new tree with the update, though the tree prior to the update may continue to exist and be used by other maps.

AvlMaps shine when (1) you need many related maps that were cloned from common ancestors and (2) cloning of keys and/or values is expensive. When you need cloned maps but cloning of entries is relatively cheap, lazy_clone_collections::BTreeMaps will often give better performance. BTrees store more entries in each node, leading to shallower trees. Updates need to clone fewer nodes, but clone more entries for each node cloned.

Implementations

impl<K, V> AvlMap<K, V>

pub fn clear(&mut self)

Drops all elements from the map.

pub fn new() -> Self

Creates a new, empty map.

Examples

use lazy_clone_collections::AvlMap;
let m: AvlMap<usize, usize> = AvlMap::new();
assert!(m.is_empty());

pub fn iter(&self) -> Iter<'_, K, V>

Creates an iterator over the map entries, sorted by key.

It is more efficient to visit each entry using for_each. (Because tree nodes may share children, the children cannot have ancestor links. The iterator tracks its progress with a Vec.) The primary advantage of iter is the functionality provided by the Iterator trait.

Examples

use lazy_clone_collections::AvlMap;

let m = AvlMap::from([(0,1), (1,2), (2, 3)]);
for (i, (k, v)) in m.iter().enumerate() {
    assert_eq!(&i, k);
    assert_eq!(&(i+1), v);
}

pub fn iter_mut(&mut self) -> IterMut<'_, K, V>

Returns iterator of the map’s entries, sorted by key, with a mutable refence to each value (and an immutable reference to each key).

The iterator obtains sole ownership of each value it returns, cloning nodes if necessary. The cloning occurs even if the calling code does not mutate the returned values. When the entire iterator is consumed, all shared nodes in self are cloned.

As with iter() and for_each, the method for_each_mut is expected to be more efficient but less ergonomic than iter_mut.

Examples

use lazy_clone_collections::AvlMap;

let mut m = AvlMap::from([(0,0), (1,1), (2,2)]);
for (k, v) in m.iter_mut() {
    *v += k;
}
assert_eq!(m.get(&0), Some(&0));
assert_eq!(m.get(&1), Some(&2));
assert_eq!(m.get(&2), Some(&4));

pub fn keys(&self) -> impl Iterator<Item = &K>

Produces an iterator over the keys of the map, in sorted order.

This is a simple projection from iter.

Examples

use lazy_clone_collections::AvlMap;

let m = AvlMap::from([(0,0), (1,1), (2,2)]);
let cnt_even_keys = m.keys().filter(|&k| k % &2 == 0).count();
assert_eq!(cnt_even_keys, 2);

pub fn values(&self) -> impl Iterator<Item = &V>

Produces an iterator over the values of the map, ordered by their associated keys.

The iterator is a simple projection from iter.

Examples

use lazy_clone_collections::AvlMap;

let m = AvlMap::from([(0,0), (1,1), (2,2)]);
let sum_values: u32 = m.values().sum();
assert_eq!(sum_values, 3);

pub fn values_mut(&mut self) -> impl Iterator<Item = &mut V>where
K: Clone,
V: Clone,

Returns an iterator of mutable references to the map’s values, ordered by their associated keys.

The iterator is a simple projection from the iterator returned by iter_mut and has the similar properties.

Examples

use lazy_clone_collections::AvlMap;

let mut m = AvlMap::from([(0,0), (1,1), (2,2)]);
for v in m.values_mut() {
    *v *= 17;
};
assert_eq!(m.get(&2), Some(&34));

pub fn for_each<F: FnMut((&K, &V))>(&self, f: F)

Applies f to each entry of the map in order of the keys.

Examples

use lazy_clone_collections::AvlMap;

let m = AvlMap::from([(0,-10), (1,0), (2,12)]);
let mut cnt_keys_gt_vals = 0;
m.for_each(|(k, v)| if k > v { cnt_keys_gt_vals += 1 });
assert_eq!(cnt_keys_gt_vals, 2);

pub fn for_each_mut<F: FnMut((&K, &mut V))>(&mut self, f: F)where
K: Clone,
V: Clone,

Applies a function to every key-value pair in the map.

The passed function must take a reference to the key type and a mutable reference to the value type. Any shared nodes in the tree are cloned, regardless of whether the contained values are mutated.

This method provides partial compensation for the absence of iter_mut().

Examples

use lazy_clone_collections::AvlMap;

let mut fmap = AvlMap::new();
fmap.insert(0, "a");
fmap.for_each_mut(|(_, v)| *v = "b");
assert_eq!(fmap.get(&0), Some(&"b"));

pub fn insert(&mut self, key: K, val: V) -> Option<V>where
K: Clone + Ord,
V: Clone,

Inserts a key-value pair in the map.

Examples

use lazy_clone_collections::AvlMap;

let mut fmap = AvlMap::new();
fmap.insert(0, "a");
assert_eq!(fmap.get(&0), Some(&"a"));

pub fn remove<Q>(&mut self, key: &Q) -> Option<V>where
K: Borrow<Q> + Clone + Ord,
Q: Ord + ?Sized,
V: Clone,

Removes a key from a map and returns the unmapped value.

Examples

use lazy_clone_collections::AvlMap;

let mut fmap = AvlMap::new();
fmap.insert(1, 2);
fmap.insert(2, 3);
assert_eq!(fmap.remove(&2), Some(3));
assert_eq!(fmap.remove(&2), None);

pub fn join_with(&mut self, key: K, val: V, rhs: AvlMap<K, V>)where
K: Clone + Ord,
V: Clone,

Join the RHS into this map.

Requires: self.last_key_value().map_or(true, |(x,)| x < key); rhs.first_key_value().map_or(true, |(y,)| key < y);

pub fn join(lhs: &Self, key: K, val: V, rhs: &Self) -> Selfwhere
K: Ord + Clone,
V: Clone,

Build a new map by joining two maps around a pivot key that divides the entries of the maps.

The constructed map contains the entries from both maps and the pivot key and the value provided for the pivot key.

Requires: lhs.last_key_value().map_or(true, |(m,)| m < key); rhs.first_key_value().map_or(true, |(n,)| key < n);

Examples

use lazy_clone_collections::AvlMap;

let f1 = AvlMap::from([(0, 'a'), (1, 'b')]);
let f2 = AvlMap::from([(3, 'd')]);
let f3 = AvlMap::join(&f1, 2, 'c', &f2);
assert_eq!(f3.get(&0), Some(&'a'));
assert_eq!(f3.get(&2), Some(&'c'));
assert_eq!(f3.get(&3), Some(&'d'));

pub fn split_off<Q>(&mut self, key: &Q) -> Selfwhere
K: Borrow<Q> + Clone + Ord,
Q: Ord + ?Sized,
V: Clone,

Moves all elements greater than or equal to the provided key into a new map and returns it.

Examples

use lazy_clone_collections::AvlMap;

let mut fmap: AvlMap<_, _> = (0..10).map(|i| (i, i * 2)).collect();
let higher_fives = fmap.split_off(&5);
assert_eq!(fmap.last_key_value(), Some((&4,&8)));
assert_eq!(fmap.get(&6), None);
assert_eq!(higher_fives.first_key_value(), Some((&5, &10)));
assert_eq!(higher_fives.get(&6), Some(&12));

pub fn diff_with(&mut self, other: Self)where
K: Clone + Ord,
V: Clone,

Removes entries with keys from the other map.

Examples

use lazy_clone_collections::AvlMap;

let mut lhs = AvlMap::from([(0,1), (1, 2)]);
let rhs = AvlMap::from([(1,5), (3,4)]);
lhs.diff_with(rhs);
assert_eq!(lhs.get(&0), Some(&1));
assert_eq!(lhs.get(&1), None);

pub fn diff(lhs: &Self, rhs: &Self) -> Selfwhere
K: Clone + Ord,
V: Clone,

Builds a map with entries from the LHS map with keys that are not in the RHS map.

Examples

use lazy_clone_collections::AvlMap;

let lhs = AvlMap::from([(0,1), (1, 2)]);
let rhs = AvlMap::from([(1,5), (3,4)]);
let d = AvlMap::diff(&lhs, &rhs);
assert_eq!(d.get(&0), Some(&1));
assert_eq!(d.get(&1), None);

pub fn sym_diff_with(&mut self, other: Self)where
K: Clone + Ord,
V: Clone,

Removes entries with keys from the other map and adds entries from the other map that have keys that are not in self.

Examples

use lazy_clone_collections::AvlMap;

let mut lhs = AvlMap::from([(0,1), (1, 2)]);
let rhs = AvlMap::from([(1,5), (3,4)]);
lhs.sym_diff_with(rhs);
assert_eq!(lhs.get(&0), Some(&1));
assert_eq!(lhs.get(&1), None);
assert_eq!(lhs.get(&3), Some(&4));

pub fn sym_diff(lhs: &Self, rhs: &Self) -> Selfwhere
K: Clone + Ord,
V: Clone,

Builds a map with entries from the LHS and RHS maps that have keys that occur in only one of the maps and not the other.

Examples

use lazy_clone_collections::AvlMap;

let lhs = AvlMap::from([(0,1), (1, 2)]);
let rhs = AvlMap::from([(1,5), (3,4)]);
let d = AvlMap::sym_diff(&lhs, &rhs);
assert_eq!(d.get(&0), Some(&1));
assert_eq!(d.get(&1), None);
assert_eq!(d.get(&3), Some(&4));

pub fn intersect_with(&mut self, other: Self)where
K: Clone + Ord,
V: Clone,

Discard entries that do not have a key from the other map.

Examples

use lazy_clone_collections::AvlMap;

let mut lhs = AvlMap::from([(0,1), (1, 2)]);
let rhs = AvlMap::from([(1,5), (3,4)]);
lhs.intersect_with(rhs);
assert_eq!(lhs.get(&0), None);
assert_eq!(lhs.get(&1), Some(&2));

pub fn intersect(lhs: &Self, rhs: &Self) -> Selfwhere
K: Clone + Ord,
V: Clone,

Creates a map with entries from the LHS that have keys in the RHS.

Examples

use lazy_clone_collections::AvlMap;

let lhs = AvlMap::from([(0,1), (1, 2)]);
let rhs = AvlMap::from([(1,5), (3,4)]);
let i = AvlMap::intersect(&lhs, &rhs);
assert_eq!(i.get(&0), None);
assert_eq!(i.get(&1), Some(&2));

pub fn union_with(&mut self, other: Self)where
K: Clone + Ord,
V: Clone,

Adds the entries from other that don’t have keys in this map.

Examples

use lazy_clone_collections::AvlMap;

let mut lhs = AvlMap::from([(0,1), (1, 2)]);
let rhs = AvlMap::from([(1,5), (3,4)]);
lhs.union_with(rhs);
assert_eq!(lhs.get(&0), Some(&1));
assert_eq!(lhs.get(&1), Some(&2));
assert_eq!(lhs.get(&3), Some(&4));

pub fn union(lhs: &Self, rhs: &Self) -> Selfwhere
K: Clone + Ord,
V: Clone,

Builds a map with entries from both maps, with entries from the LHS taking precedence when a key appears in both maps.

Examples

use lazy_clone_collections::AvlMap;

let mut lhs = AvlMap::from([(0,1), (1, 2)]);
let rhs = AvlMap::from([(1,5), (3,4)]);
lhs.union_with(rhs);
assert_eq!(lhs.get(&0), Some(&1));
assert_eq!(lhs.get(&1), Some(&2));
assert_eq!(lhs.get(&3), Some(&4));

pub fn first_key_value(&self) -> Option<(&K, &V)>

Returns the key-value pair for the least key in the map

Examples

use lazy_clone_collections::AvlMap;

let fmap = AvlMap::from([(2,0), (1,0)]);
assert_eq!(fmap.first_key_value(), Some((&1, &0)));

pub fn last_key_value(&self) -> Option<(&K, &V)>

Returns the key-value pair for the greatest key in the map

Examples

use lazy_clone_collections::AvlMap;

let fmap = AvlMap::from([(2,0), (1,0)]);
assert_eq!(fmap.last_key_value(), Some((&2, &0)));

pub fn contains<Q>(&self, key: &Q) -> boolwhere
K: Borrow<Q>,
Q: Ord + ?Sized,

Tests if self contains an entry for the given key.

pub fn get<Q>(&self, k: &Q) -> Option<&V>where
K: Borrow<Q>,
Q: Ord + ?Sized,

Returns a reference to the value associated with k.

Example

use lazy_clone_collections::AvlMap;

let mut fmap = AvlMap::new();
fmap.insert(0, 100);

assert_eq!(fmap.get(&0), Some(&100));

pub fn get_mut<Q>(&mut self, k: &Q) -> Option<&mut V>where
K: Borrow<Q> + Clone + Ord,
Q: Ord + ?Sized,
V: Clone,

Returns a mutable reference to the value associated with k.

Example

use lazy_clone_collections::AvlMap;

let mut fmap = AvlMap::new();
fmap.insert(1, 7);

*fmap.get_mut(&1).unwrap() = 2;
assert_eq!(fmap.get(&1), Some(&2));

pub fn entry(&mut self, key: K) -> Entry<'_, K, V>where
K: Clone + Ord,
V: Clone,

Returns an Entry that simplifies some update operations.

pub fn is_empty(&self) -> bool

Returns true if self contains no entries, false otherwise.

pub fn len(&self) -> usize

Returns the number of entries in self.