Struct lazy_clone_collections::AvlSet

pub struct AvlSet<T> { /* private fields */ }

A sorted set of values.

The implementation is a wrapper around [AvlMap<T,()>].

Implementations

impl<T> AvlSet<T>

pub fn append(&mut self, other: &mut Self)where
T: Ord + Clone,

Moves all elements from other into self and leaves other empty.

pub fn clear(&mut self)

Removes all the entries from self.

pub fn contains<Q>(&self, value: &Q) -> boolwhere
T: Borrow<Q>,
Q: Ord + ?Sized,

Tests if self contains the given value.

pub fn difference<'a>(&'a self, other: &'a Self) -> Difference<'a, T>where
T: Ord,

Returns an iterator over elements in self and not in other

pub fn first(&self) -> Option<&T>

Returns the least value in the set.

pub fn get<Q>(&self, value: &Q) -> Option<&T>where
T: Borrow<Q>,
Q: Ord + ?Sized,

Returns a reference to the element matching value, if it exists

pub fn insert(&mut self, value: T) -> boolwhere
T: Clone + Ord,

Inserts the given value and returns true if self did not already have the value and returns false otherwise.

pub fn intersection<'a>(&'a self, other: &'a Self) -> Intersection<'a, T>where
T: Ord,

Returns an iterator of the values that are in both self and other.

pub fn is_disjoint(&self, other: &Self) -> boolwhere
T: Ord,

Returns true if self and other have no common values and false otherwise

pub fn is_empty(&self) -> bool

Returns true if self is the empty set, false otherwise.

pub fn is_subset(&self, other: &Self) -> boolwhere
T: Ord,

Tests if self is a subset of other.

pub fn is_superset(&self, other: &Self) -> boolwhere
T: Ord,

tests if self is a superset of other.

pub fn iter(&self) -> Iter<'_, T>

Returns an iterator over self’s values in sorted order.

pub fn last(&self) -> Option<&T>

Returns the greatest value in self.

pub fn len(&self) -> usize

Returns the number of elements in self.

pub fn new() -> Self

Returns a new, empty set.

pub fn new_diff(lhs: Self, rhs: Self) -> Selfwhere
T: Clone + Ord,

Creates a new set with the elements of lhs that are not in rhs.

pub fn new_intersection(lhs: Self, rhs: Self) -> Selfwhere
T: Clone + Ord,

Creates a new set with the elements of lhs that are also in rhs.

pub fn new_union(lhs: Self, rhs: Self) -> Selfwhere
T: Clone + Ord,

Creates a new set with the elements of both lhs and rhs.

pub fn new_sym_diff(lhs: Self, rhs: Self) -> Selfwhere
T: Clone + Ord,

Creates a new set with the elements that are in lhs or rhs but not both.

pub fn pop_first(&mut self) -> Option<T>where
T: Clone,

Removes the first element from the set and returns it.

pub fn pop_last(&mut self) -> Option<T>where
T: Clone,

Removes the last element from the set and returns it.

pub fn remove<Q>(&mut self, value: &Q) -> boolwhere
T: Borrow<Q> + Clone + Ord,
Q: Ord + ?Sized,

Removes the given value from self returning true if the value was present and false otherwise.

pub fn replace(&mut self, value: T) -> Option<T>where
T: Clone + Ord,

Replace and return the matching value in the map.

pub fn retain<F>(&mut self, f: F)where
T: Clone + Ord,
F: FnMut(&T) -> bool,

Retain values for which f returns true and discard others

pub fn split_off<Q>(&mut self, key: &Q) -> Selfwhere
T: Borrow<Q> + Clone + Ord,
Q: Ord + ?Sized,

Removes all elements greater or equal to key and returns them.

pub fn symmetric_difference<'a>(
&'a self,
other: &'a Self
) -> SymmetricDifference<'a, T>

Returns an iterator over elements in self or other but not both.

pub fn take<Q>(&mut self, value: &Q) -> Option<T>where
T: Borrow<Q> + Clone + Ord,
Q: Ord + ?Sized,

Removes and returns the set member that matches value.

Examples

use lazy_clone_collections::AvlSet;

let mut s = AvlSet::new();
s.insert("abc".to_string());
s.insert("def".to_string());
assert_eq!(s.take("abc"), Some(String::from("abc")));
assert_eq!(s.len(), 1);

pub fn union<'a>(&'a self, other: &'a Self) -> Union<'a, T>

Returns an iterator over the elements of self and other, ordered by key.

Common elements are only returned once.