Struct lazy_clone_collections::Stack

pub struct Stack<T> { /* private fields */ }

Implements a stack with constant time push, pop, and clone operations.

The stored elements must implement Clone. During pop operations, popped items are moved if they are not shared among different stacks and cloned if they are.

Examples

In this example, we create a Stack of Boxes to illustrate Stack’s use of clone on stored elements.

use lazy_clone_collections::Stack;

let mut s = Stack::new();
s.push(Box::new(0));
let mut t = s.clone();

// At this point, s and t share their internal storage.

s.push(Box::new(1));
t.push(Box::new(2));

// s and t still have common storage of the Box(0) element, but also
// their own storage for Box(1) and Box(2), respectively.

assert_eq!(s.pop(), Some(Box::new(1))); // pop moves its Box(1) (unshared)
assert_eq!(s.pop(), Some(Box::new(0))); // pop clones Box(0) (shared with t)
assert_eq!(t.pop(), Some(Box::new(2))); // pop moves Box(2) (unshared)
assert_eq!(t.pop(), Some(Box::new(0))); // pop moves Box(0) (now unshared)

Implementations

impl<T: Clone> Stack<T>

pub fn new() -> Self

Creates an empty stack.

Examples

use lazy_clone_collections::Stack;

let stack: Stack<i32> = Stack::new();

pub fn iter(&self) -> Iter<'_, T>

Creates an iterator from the top to the bottom elements of the stack.

Example

As with most collections, you cannot modify a Stack while you are iterating over it. However, because cloning a collection is so cheap, it is sometimes viable to iterate over a clone while you modify the original. In this example, we iterate over a clone while moving the odd elements in the original stack to its top.

use lazy_clone_collections::Stack;

let mut s: Stack<_> = (0..6).collect();

for (i, n) in s.clone().iter().enumerate() {
    if n % 2 == 1 {
        // remove indexes bottom to top, while iter goes top to bottom
        let x: i32 = s.remove(5 - i);
        assert_eq!(x, *n);
        s.push(x);
    }
}

assert_eq!(s.pop(), Some(1));
assert_eq!(s.pop(), Some(3));
assert_eq!(s.pop(), Some(5));
assert_eq!(s.pop(), Some(4));
assert_eq!(s.pop(), Some(2));
assert_eq!(s.pop(), Some(0));
assert_eq!(s.pop(), None);

pub fn iter_mut(&mut self) -> IterMut<'_, T>

Returns an iterator with mutable references.

Before iterating to a shared node, the iterator will clone it.

Examples

use lazy_clone_collections::Stack;

let mut s = Stack::new();

s.push(0);
let mut t = s.clone();
s.push(1);
t.push(2);

for x in s.iter_mut() {
    *x += 100;
}

assert_eq!(s.pop(), Some(101));
assert_eq!(s.pop(), Some(100));
assert_eq!(s.pop(), None);

assert_eq!(t.pop(), Some(2));
assert_eq!(t.pop(), Some(0));
assert_eq!(s.pop(), None);

pub fn push(&mut self, val: T)

Pushes an element on top of the stack.

Examples

use lazy_clone_collections::Stack;

let mut s = Stack::new();

s.push("hello");
assert_eq!(s.top(), Some(&"hello"));

pub fn top(&self) -> Option<&T>

Returns a reference to the top of the stack, or None if the stack is empty.

Examples

use lazy_clone_collections::Stack;

let mut s: Stack<_> = (0..3).collect();
assert_eq!(s.top(), Some(&2));

s.clear();
assert_eq!(s.top(), None);

pub fn top_mut(&mut self) -> Option<&mut T>

Returns a mutable ref to the top of the stack or None if empty.

Clones the top node if it is shared.

Examples

use lazy_clone_collections::Stack;

let mut s: Stack<_> = (0..=3).collect();

assert_eq!(s.top_mut(), Some(&mut 3));

*s.top_mut().unwrap() += 5;

assert_eq!(s.pop(), Some(8));
assert_eq!(s.pop(), Some(2));

pub fn pop(&mut self) -> Option<T>

Returns the top of the stack or None if empty.

pub fn len(&self) -> usize

Returns the number of elements in the stack.

Examples

use lazy_clone_collections::Stack;

let mut s: Stack<_> = (0..5).collect();
assert_eq!(s.len(), 5);

pub fn clear(&mut self)

Removes all elements from the stack.

Examples

use lazy_clone_collections::Stack;

let mut s: Stack<_> = (0..3).collect();
assert_eq!(s.len(), 3);
s.clear();
assert!(s.is_empty());

pub fn contains(&self, x: &T) -> boolwhere
T: PartialEq<T>,

Tests if the element x occurs in the stack.

Examples

use lazy_clone_collections::Stack;

let s: Stack<_> = (3..8).collect();
assert!(s.contains(&4));
assert!(!s.contains(&0));

pub fn is_empty(&self) -> bool

Tests if there are any elements in the stack.

Examples

use lazy_clone_collections::Stack;

let mut s = Stack::new();
assert!(s.is_empty());
s.push(47);
assert!(!s.is_empty());

pub fn remove(&mut self, at: usize) -> T

Removes and returns the element at the given index.

The bottom element of the stack has index 0 and the top has index self.len() - 1. Any shared nodes above the dropped element are cloned.

Panics

Panics if the index is greater than or equal to the length of the stack.

Example

use lazy_clone_collections::Stack;
let mut s = Stack::from(vec!['a','b','c', 'd']);
assert_eq!(s.remove(1), 'b');
assert_eq!(s.len(), 3);
assert_eq!(s.pop(), Some('d'));
assert_eq!(s.pop(), Some('c'));
assert_eq!(s.pop(), Some('a'));
assert_eq!(s.pop(), None);

pub fn split_off(&mut self, at: usize) -> Self

Splits the stack at the given index, retaining the bottom elements from [0..at] and returning a new stack with the top elements from [at..].

The bottom element has index 0 and the top has index self.len()-1. Any shared nodes in the returned top portion of the stack are cloned.

Panics

Panics if at > self.len().

Examples

use lazy_clone_collections::Stack;

let mut s: Stack<_> = (0..21).collect();
let t = s.split_off(7);

assert_eq!(s.len(), 7);
assert_eq!(t.len(), 14);
assert!(s.into_iter().cmp((0..7).rev()).is_eq());
assert!(t.into_iter().cmp((7..21).rev()).is_eq());